Update Exit.cs
Made the trigger for tags which is a convention and takes one less line of code. 

Deleted the "you win", the wait for seconds only works in an IEnumerate functions which has its own syntax. 

Added a public string that allows us to change the next level we want to load, making this script use able in all of our scenes.